+++ src/application/Randomizer.h
@@ -31,6 +31,8 @@ public:
     std::map<int, LocationData> GetCheckData() const;
     void OnConnected();
     void OnGameCompleted();
+    void ShowStatusInformation(std::string information);
+    void QueueNewMessage(std::string information);
 
 private:
     DisplayManager& _displayManager;