+++ src/pch.h
@@ -14,4 +14,7 @@
 #include "../lib/sadx-mod-loader/mod-loader-common/ModLoaderCommon/IniFile.hpp"
 #include "../lib/APCpp/Archipelago.h"
 
+
+
 #endif //PCH_H
+#define FunctionHookAdd(address, hookFunction) FunctionHook<void> hook_##address(address, [] { hookFunction(); hook_##address.Original();  })